InnerSymDCE: Fix to handle ports, fields, and extmodule's.
Also fix the statistics.

The InnerRef walk can be done in parallel, future work.

Also add dropSymbol method for removing a particular symbol,
this was previously added to IST but requires more support in
HW and related dialects, keep this local for now since this
is only supported in FIRRTL presently anyway.