[FIRRTL] Change LowerXMR to create XMRRefOps
Change the LowerXMR pass to emit XMRRefOps instead of VerbatimExprOps.
This is done to enable later inlining of these operations into their
usage sites, e.g., directly into a bind.

As this has to bridge between FIRRTL and SV, this uses an unrealized
conversion cast to enable usage of an InOut type from FIRRTL.  This is
already correctly handled by LowerToHW.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>