[Handshake] Cleanup Handshake operations builders (#4416)
* isControl paramater for some HandshakeOps builders

Logic to determine whether a Handshake operation is control-only
or not was moved outside the builders and into the
StandardToHandshake conversion pass. Modified builders now take an
isControl parameter.

* Cleaned up merge-like operations builders

This commit refactors and cleans up the merge insertion logic in
StandardToHandshake. The pass now creates valid merge-like operations
from the beginning with backedges as operands. All backedges are then
resolved in reconnectMergeOps. This allows us to remove some operation
builders that were only used during StandardToHandshake.

* Handshake builders cleanup and SOST interface

This commit removes a lot of custom builders for Handshake operations,
and re-enables default builders for most operations. The commit also
replaces the sost namespace (along with its operation attributes logic)
with an SOSTInterface that Handshake operations can implement to achieve
the same purpose with less code and more reliability. This in turns
facilitates the deletion of some operation attributes that were
redundant.

A lot of tests still fail on this commit due to largely unupdated
parse and print logic for Handshake operations.

* Fixed many issues, all tests passing

This commit fixes many issues introduced during Handshake operation
builders cleanup, many related to parsing, printing, or constructing
operations. In the HandshakeToFIRRTL pass, the control-ness of each
Handshake operation is cached in an attribute before lowering to
accomodate the new logic to determine whether an operation is
control-only.

* Verification and type inference for HandshakeOps

This commit adds some verification logic as well as return type
inference for a number of Handshake operations. The latter allows to
shorten the creation of Handshake operations in many places as the
return type(s) need no longer be specified at creation.

The commit also introduces a simple builder for ForkOp and LazyForkOp.