[HW] Make NamedRefAttr get-able/print no type
Change the ODS description of NameRefAttr to make NamedRefAttr capable
of generating a "getter".  This requires defining its return type.
Additionally, define this as having "NoneType" to work around assembly
format restrictions that this cannot come before a ":" in the assembly
format even though it fundamentally has no ":" in its serialized type
representation.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>