[hlstool] Remove header to fix flaky build
The `InitAllPasses.h` header file includes all the generated pass header
files, e.g. `"circt/Dialect/SystemC/Passes.h.inc"`.  To make sure this
header file is generated properly before we try to include it, we have
to depend on the associated transform library.  Since hlstool does not
depend on every single transform library, this causes the build to
occasionally fail.  To fix this, we need to only include header files
belonging to the dialects we are actually depending on.

This error can be seen in the following build:
https://github.com/llvm/circt/actions/runs/3657570236/jobs/6181309548