Avoid building `circt-lec` if the detected `Z3` version is outdated
Added logic to conditionally build the tool only when detecting a Z3
installation newer than version 4.8.11. Previous versions do not provide
the `sge` and `sgt` operators in the C++ API, which are instead referenced
by the cirt-lec codebase and would cause linting and compiling issues.
Discussed with @maerhart and @teqdruid.