[FIRRTL] Simplify getFieldName utilities
Combine two getFieldName utility functions for computing a
human-readable name (a string) from a value.  Previously, this consisted
of one that took a "bool &" and one that did not.  Change this to a
single function that returns a pair of a string and a bool.  Fixup the
rest of the codebase to use the new function.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>