[FIRRTL] Add nameSafe option to getFieldName
Make the getFieldName utility usable to generate IR names.  Previously,
this utility was primarily intended to generate error messages from the
IR.  However, it is also desirable to generate names for in-IR things
while avoiding weirder characters like ".", "[", and "]". This adds a
boolean flag to getFieldName that will convert these to saner names.
E.g., what would previously be:

    foo[42].bar

Can now be optionally generated as:

    foo_42_bar

The former is great for an error message, while the latter is better for
a generated port name.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>