[CI] uploadBinaries: allow workflow_dispatch, test what we publish, drop cache (#4474)
Enable the binary workflow to be run manually, but only upload binaries if it's a tag. I am not sure what happens if they're already present, so probably best to use this wisely.

While visiting, run the basic lit tests for LLVM/MLIR before bundling them for others (and sometimes ourselves) to use.

Drop cache as well: partly to avoid increased pressure on the cache from binaries/components built to run the tests that we don't presently, but also because I'm not sure the cached LLVM will likely ever be a cache hit since we release and update LLVM at about the same frequency.  In the future, this may be useful if we did do more frequent releases, or ever promoted a pre-release to a an actual one.

cc #4473.