HWModuleLike: setPortSymbolAttr, related.
Presently non-FIRRTL HWModuleLike's, at least in the HWModuleLike
interface methods, do not assume symbols for a given argument
are stored under any particular name and search for the first InnerSymAttr.

Without adding a new such assumption, this means we can't really
add a new InnerSymAttr if one doesn't exist.

For now error out if that's attempted, this is enough to update or remove
existing symbols when present using the same search strategy.