Remove intermediate StdToHandshake conversion ops
This commit removes the StartOp, EndOp, and TerminatorOp from the
entire codebase, since they were used exclusively as intermediate
operations during std-to-handshake conversion and have caused issues in
the past. StartOp's purpose is achieved through a new field (startCtrl)
in the HandshakeLowering class, while TerminatorOp and EndOp's purpose
is obtained by keeping cf-level terminators in the IR for longer than
before.

Unit tests on the conversion pass are modified manually to reflect the
fact that an additional MergeOp is created by the pass for the
control-only network in the entry block of each function (this has no
impact on circuit behavior, and can be canonicalized away).

The commit also includes some more detailed documentation for the
merge-insertion step of the conversion pass.