[FIRRTL] Generate docs for each ODS file
Change the way FIRRTL ODS documentation is generated to manually run
each file through mlir-tblgen.  This is primarily done to enable more
cleanly adding documentation about OpInterfaces and to remove
information about type constraints (which are ephemeral and do not
appear in the generated C++).

Note that this currently produces buggy Markdown for OpInterfaces with a
defined function body.  This will be fixed by:

  - https://reviews.llvm.org/D140590

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>