[Seq] Create new clock enabled register op
Introduces `compreg.ce` which is identical to the existing `compreg` but
includes a `clockEnable` signal to enable/disable reading the input.
This op makes it easier to create SV with the behavioral syntax rather
than the structural mux that we currently construct. Could be lowered
into the structural mux if that is desired.