[firtool] Run SymbolDCE as late as possible. (#4504) (#4507)
This pass should generally be run as late as possible so it can clean
up after all prior passes. It currently runs just before the
BlackBoxReader pass, which happens to leave behind some dead symbols.

This moves it as late as possible: just before InnerSymbolDCE. While
there is no new integration test of the firtool pipeline with this
change, it has been shown to remove symbols that would otherwise
remain.