[Seq] Replace unitialized array element with constant zero
For reset-less 1d array registers, replace an unitialized elemenent with
constant zero. For example, let `r` be a 2xi1 register and its next value
be `{foo, r[0]}`. `r[0]` is connected to itself so will never be
initialized. If we don't enable aggregate preservation, `r_0` is relpaced
with `0`. Hence this canonicalization replaces 0th element of the next
value with zero to match the behaivor.