[PyCDE] Add `ControlReg` construct
A 'control register' is a single bit register which gets set and reset
based on pulses. Asserts are signals which causes the output to go high
(on the next cycle). Resets do the opposite. If both an assert and a
reset are active on the same cycle, the assert takes priority.