[Utils] Bump OR-Tools version.
Simple version bump to keep things up-to-date. The SCIP solver was recently re-licensed under Apache 2.0, however, the version OR-Tools pulls in is unfortunately still subject to the old license, so no changes there.

I tested locally that no further changes are required to CIRCT. A pull request to update the integration test image to the same version has been created.