[FIRRTL] Use circuit location for annotation file parse errors.
The inline annotation location is invalid if none are specified,
and since the error isn't in inline annotations just point to circuit.

Would be nice to point into the file itself, or mention it,
but that will require reworking how these files are loaded/passed around.

For now, let's not crash.
This is the same as we do for OMIR annotation files and their error reporting.

Fixes #4538.