Tests for maximize-SSA pass and pass fix
This commit adds unit tests for the maximize-SSA pass, and fixes a bug
in the pass itself. The latter happened when a conditional branch had
its two branches pointing to the same block. In that case, potential new
block arguments would incorrectly be added twice to the list of branch
arguments (due to the fact that getPredecessors() doesn't return a list
of *unique* predecessors in that case), causing a pass failure.

This commit also removes some redundant comments.