[FIRRTL] No final inst in ExtractInstances output
Change the ExtractInstances pass to no longer include the final
extracted instance in the output file it optionally generates.

The new format of this file has the _path to the module_ where the
extracted instance used to be located as opposed to the _path to the
pre-extracted instance_ with a name that is either the new name of the
instance (MFC, previously) or the old instance name (SFC).

Concretely, this changes the output from something like (MFC,
previously):

    clock_gate_0 -> Foo.bar.baz.gate_0
    clock_gate_1 -> Foo.bar.gate_1

To:

    clock_gate_0 -> Foo.bar.baz
    clock_gate_1 -> Foo.bar

Here, "clock_gate_<n>" is the name of the ports that a user is supposed
to look at and "gate_<n>" is the name of the extracted instance outside
of the module that contains these ports.  This is weird because it leaks
information that is not supposed to be visible.  Specifically, this
leaks that there is an instance called "gate_<n>" that is outside the
design and not in "Foo.bar.baz" or "Foo.bar".  This post-extracted
instance name is necessarily uniqued within the module where it was
extracted ("gate_0" and "gate_1" could be generated if multiple
instances with the same name, "gate",are extracted).

The Scala-based FIRRTL Compiler (SFC) does something different, but also
wrong.  The SFC will emit the above like:

    clock_gate_0 -> Foo.bar.baz.gate
    clock_gate_1 -> Foo.bar.gate

Here, "gate" is the original name of the instance.  However, "gate",
like "gate_<n>", does not match the name of anything in either
"Foo.bar.baz" or "Foo.bar" as this instance was extracted.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>