[FIRRTL] Always create ExtractSeqMems extract file
Change the ExtractInstances pass to always create an ExtractSeqMems file
if the appropriate annotation is passed in.  This is done to align with
the SFC implementation of this pass (and to pass an internal unit test).

This uses the same strategy as for CreateSiFiveMetadata where a dummy
verbatim is always created.  This simplifies later logic.  This could be
improved in the future to only generate the dummy verbatim when
necessary (which would remove a now mandatory 1-newline offset that is
added to every ExtractSeqMems file created.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>