[FIRRTL] Add GrandCentral no-bind option
Add an option to the GrandCentral pass that will _not_ bind in the
companion module and will _not_ emit the SV interface instantiation.
This is done to enable GrandCentral collateral to be synthesizable.
This is motivated by making all the assertions that are frequently
crammed into a Grand Central companion as eligible for synthesis.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>