[FIRRTL][GC] Always generate the scope yaml file
When we added the flag to directly instantiate companions, we also
stopped generating interfaces.  This flag skipped generating the
interface metadata file, which caused issues for flows which still
expected it to be generated.  This commit changes the logic to skip
adding interfaces to the worklist instead of an early break, which
causes us to still generate the metadata, but with no entries.