[PyCDE][NFC] Create signals from Python objects through Types
Instead of calling `_obj_to_value` from the support library, use some
OOP to leverage the `Type` system to do it. The `_obj_to_value` function
dates back to when we were using raw `ir.Types` and not extending them.
Now that we have a proper type hierarchy (and can rely on it being
used), this change is possible.