Migrate to the new folding api
In the old folding API, folders would get an array of attributes, where each
entry corresponds to an input operand of the op.

In the new folding API, this array is wrapped by an op-specific FoldAdaptor,
which gives us structured and nicely named accessors to the underlying array of
attributes.

This new folding API will soon be the only way to write folders, and as such,
ODS generates a warning when a dialect has not switched over to the new folding
API.

This commit moves every dialect in circt to the new folding API.