[FIRRTL] Chisel Memory File Init Integration Tests
Add an integration test of Chisel features for loading memories:

  1. loadMemoryFromFile
  2. loadMemoryFromFileInline

These APIs provide a way for a user to specify that a memory should get
either a "$readmemh" or "$readmemb" in an initial block using a
specified file.  For (1), the memory is initialized using a module that
is bound into the memory module and does its "$readmem*" using an XMR.
For (2), the memory is initialized directly inside the memory module.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>