[firtool] Use CIRCT's StripDebInfo pass instead of upstream
With the recent change to how port location information is handled, the
upstream pass StripDebugInfo no longer leaves the IR in a valid state,
as it does not update the debug information stored in module ops.  To
fix this problem, we just have to use the local version of this pass,
StripDebugInfoWithPred and tell it to remove all locations.