[ESI] Introduce ESI pure module op
A module containing only ESI channels and modules with only ESI ports.
All non-local connectivity is done through ESI services. If this module
is the top level in the design, then the design's actual top level ports
are defined by a BSP.

Useful on its own for simulation and BSPs which don't define a
top-level.