[CheckCombCycles] Iterative Tarjan's SCC to detect cycles (#4642)
Re-implement the pass using an iterative Tarjan's SCC algorithm instead of the
current recursive implementation. Should handle aggregates and subaccess op.
Overview of the algorithm:
Process each module in the InstanceGraph by traversing it bottom up.
For each module, build a map of all the FieldRefs that each Value
can refer to, and the map of all the Values that refer to a FieldRef.
Given any Value, we can compute the set of Values that alias with it,
that is refer to the same FieldRef.
Now perform a DFS over the "interesting" ops in the module.
The set of Values whose successors are currently being visited are recorded
in the Visiting set. Once all the successors of a Value are visited, it is
 removed from the Visiting set. If a Value is visited and it already exists
 in the Visiting set, then a Cycle exists in the graph.
 The Combinational paths of a module are inlined at the instance, to detect
 combinational cycles across modules.