[FIRRTL][WiringProblem] Support DataTap+LegacyWiring.
More specifically, support wiring to ports of an instance
used for data taps w/internal paths.

Insert an IR constrcut (unrealized_conversion_cast) as placeholder
to ensure a stable reference to the target value.

This is the same approach DataTaps already uses to handle
the problem.

Test[1] now passes and is valgrind-clean.
(valgrind or ASAN complain if only the error is removed)

[1] test/Dialect/FIRRTL/legacy-wiring-and-data-taps.fir