[Arc] Add variadic op simplification pass
Add the `SimplifyVariadicOps` pass which takes the common combinational
operations and converts them from their variadic representation into a
chain of binary operations. The pass reorders the variadic operands such
that they appear in the order in which their definitions appear in the
basic block. The binary operations are then inserted as soon as both
their operands become available. This distributes the operations through
the block and reduces the lifetimes of the original op's operands.