[Arc] Add state lowering pass
Add the `LowerState` pass and accompanying operations to convert a
design from a pure state transfer graph composed of arcs to a more
procedural read-modify-write representation. After this transformation
the design is a significant step down the path of becoming a software
model.

The `LowerState` pass proceeds as follows:

- Group all `arc.state` ops according to their clock into
  `arc.clock_tree`s. Operations that are on direct input-to-output
  passthrough or on state-to-output paths are grouped into a
  `arc.passthrough` op.
- Introduce an explicit state storage allocation op for every state op,
  memory op, as well as every input and output port of the root module.
- Replace the root `hw.module` with a `arc.model` which no longer has
  any input and output ports, but a storage pointer argument instead.
  Storage allocation ops represent specific chunks of memory behind this
  pointer.
- Split every `arc.state` op with latency >0 up into a `arc.state_read`,
  `arc.state` with latency 0, and `arc.state_write` operation. This
  essentially breaks `arc.state` up into two parts: a read for all users
  of the state, and a transfer function plus write for all operands of
  the arc.

In doing so, the sea-of-gates representation of the HW dialect, which
is a pure graph without op ordering constraints, is converted into a
a sea-of-clocks, where each group contains the parts of the circuit that
are triggered by the same clock. The actual computation that occurs on
that trigger is represented procedurally in a proper SSA/CFG block.

TL;DR: This goes from "How do the gates connect together?" to "How does
a computer simulate these gates?".

Co-authored-by: Martin Erhart <maerhart@outlook.com>
Co-authored-by: Zachary Yedidia <zyedidia@gmail.com>