[Arc] Add simple runtime helpers
Add utilities to allow C++ to link against and drive a software model
generated from the arc dialect.

The `arcilator-header-cpp.py` utility accepts the state JSON file
generated by the `PrintStateInfo` pass and converts it into a C++ header
that makes the lowered LLVM model feel close to Verilator output. The
mapping is not perfect: there is no `eval` function, but dedicated clock
and passthrough functions to be called instead. The commonalities of the
generated C++ header are collected in `arcilator-runtime.h`.

The utilities and headers aren't installed yet properly.

Co-authored-by: Martin Erhart <maerhart@outlook.com>
Co-authored-by: Zachary Yedidia <zyedidia@gmail.com>