[Arc] Add pass to make resets and enables explicit in the state ops
A pass to detect resets and enables in arcs. Making them explicit allows
another pass to pick them up more easily and, e.g., group the ones with
the same conditions together to reduce the number of branches, position
the ones with the same resets next to each other in memory to do a
simple memset over a continuous region in mem, etc.

I don't particularly like the way this is represented in the IR right
now because it makes the state ops more complex and basically everything
is just represented with this single op. It also forces us to have a
static order of reset and enable (currently the reset operand always has
higher precedence).