[CI] Add a workflow for building and uploading Python wheels.
The workflow uses cibuildwheel to build the wheels, and uploads them
to GitHub. They're added to the artifacts for manual runs, and added
to the release assets when triggered by a release.

Currently, manylinux wheels are built for CPython 3.7 and 3.10.

This also updates the packaging scripts to include setuptools_scm to
take version numbers from Git tags.