Remove pipelining from simplified single-address memory reads
My initial understanding of memory latencies was incorrect and I wrongly inserted pipelining after the resulting register.
However, it is the address and enable signals that must be pipelined. For these memories, the address is always 0 and the
enable signal can be discarded, allowing the read ports to be folded to the resulting register.

Fixed #4734