[ExportVerilog] Add KEEP attr to Vivado array index bug workaround
Rename the `disallowArrayIndexInlining` lowering option to
`mitigateVivadoArrayIndexConstPropBug` to more clearly convey the fact
that this is intended as a Vivado bug workaround.

Since this is now Vivado-specific, add the `(* keep = "true" *)` SV
attribute to the array index wire, which is necessary to properly
mitigate the Vivado synthesis issue we were observing.