[FIRRTL][NFCI] Use FConnectLike to abstract over connection types.
Ensure code continues to work for anything that's FConnectLike,
without having to add it to each of these locations.

Code with behavior very specific to ConnectOp and/or StrictConnectOp are left
unmodified, but if all that matters is source => destination then use
FConnectLike.

Prefer `getDest()` (and `getSrc()`) over hardcoding operand numbers.