[FIRRTL] Check subindex in-bounds via constraint
Change the way that a subindex's index is validated by adding an
explicit multi-entity constraint on the operation.  Previously, this was
handled via an error in the implementation of its InferTypeOpInterface
method.  This was an abuse of the InferTypeOpInterface because the op
should be verified before this and any error during type inference
_should_ result in a crash.  However, this is actually a standard
invalid IR error.  Remove this check in type inference.

Fixes #4753.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>