[SVExtractTestCode] Process modules in "bottom up" order.
Collect modules according to the hierarchy of the instance
graph. Modules are sorted so that modules whose instances are higher
in the hierarchy come first. This allows us to process modules "bottom
up", so test code is extracted from children before parents. That
allows the parents to potentially extract test code from their children.