[Arc] Relax arc body ops to isMemoryEffectFree
Instead of checking for `isPure` on arc body ops, which includes a check
for `isSpeculatable` which operations like `scf.if` does not pass, only
check for `isMemoryEffectFree`. Also only perform the check on the
top-level operations, since side-effecting ops with regions implement
the `HasRecursiveMemoryEffects` trait that causes the check to recur on
its own.

Add a check that arc bodies can contain `scf.if`.