[Support] Add `ImplicitSSAName` directive for string attributes
Move the `{parse,print}ImplicitSSAName` directive currently used by the
SystemC dialect into the support library and make the SV dialect use it
as well. This directive works on a `StringAttr` instead of the entire
attribute dictionary of the operation, which has a few advantages:

- Name is an explicit ODS attribute on the operation
- Name gets automatically elided in attr-dict printing (TableGen elides
  all ODS-defined attributes)
- The fact that the "name" attribute has a specific meaning is exposed
  in ODS more explicitly, and the assembly format makes this more
  apparent by having a dedicated `name <ATTR>` portion.

NFC besides the change in assembly format.