[HW] Add hw.wire operation
Add the `hw.wire` operation to the HW dialect. A wire offers the ability
to assign a human-readable name or an inner symbol to an edge in the SSA
graph, for human consumption or remote references through a symbol.

This fills a representational gap where language frontends like FIRRTL
have to resort to `sv.wire` if they need any of the naming or symbol
reference capabilities of a wire, but in doing so they also have to buy
into the SV-specific quirks and connect semantics of `sv.wire`, which
are annoying to analyze. Almost all uses in FIRRTL don't care about the
connect semantics though. The new `hw.wire` op allows for expressing a
wire without the connect semantics, leveraging the graph-ness of the HW
dialect, which makes analysis a lot easier and will allow frontends to
significantly reduce the number of SV-specific ops they emit. Decoupling
frontends from the SV emission backend is very valuable in practice, as
it allows for the intermediate HW representation to be consumed by other
tools like simulators without them having to try and implement the SV
semantics of a wire.

In a nutshell, this allows for the following replacement in almost all
cases:

```mlir
// before
%myWire = sv.wire sym @mySym : !hw.inout<i42>
%0 = sv.read_inout %myWire : !hw.inout<i42>
sv.assign %myWire, %1 : i42
hw.output %0 : i42

// after
%myWire = hw.wire %1 sym @mySym : i42
hw.output %myWire : i42
```

This commit adds the `hw.wire` op, canonicalizations, and tests for it.
It also adds `BuilderUtils` with a helper to accept `StringAttr`,
`StringRef`, or `Twine` in a single builder argument, which is going to
be useful for other operations as well. The changes to ExportVerilog are
limited to qualifying `WireOp` as `sv::WireOp` to disambiguate from the
new `hw::WireOp`.