[reduce] Fix crash on nested ops; fix poor scaling of module op counting (#4860)
Fix a source of crash in `circt-reduce` where the tool would try to
apply a reduction to a parent operation *and* some of its nested child
operations at the same time. Add a set to keep track of which operations
were already affected by a reduction on their parent, and skip those.

Also fix poor scaling behaviour in `computeTransitiveModuleSize` which
made the FIRRTL module externalizer and instance stubber very slow. The
new simplified implementation doesn't produce reduction estimates as
accurately, but it is a lot faster and seems sufficient in most cases.

Implements parts of #4100.