[FIRRTL][InferWidths] Infer InvalidValueOp to zero width by default
The `InferWidths` pass currently errors out if the input contains any
`InvalidValueOp`s that aren't used as the right hand side of a connect.
Make the pass more forgiving by inferring these ops to zero width unless
they are used in a connect, such that the pass doesn't crash as easily
on reduction inputs.

NFC for all but a few very obscure inputs.

Fixes #4859.