[FIRRTL][FIRParser] rwprobe() parsing using Forceable.
Emit "forceable" declarations as forceable, so they have a rwprobe
reference to use later if needed.

(don't mutate IR during parsing)

After parsing a module, walk and demote declarations that weren't
used in a `rwprobe()`.  This would also happen with the canonicalizer,
but don't want this even in the `--parse-only` output or seen
by any of the early passes.