[FIRRTLToHW] Create most temporary wires as `hw.wire` (#4828)
Add a `createTmpHWWireOp` similar to `createTmpWireOp` that creates a
throw-away `hw.wire` which is guaranteed to be deleted after the
lowering, instead of an `sv.wire` that may be deleted. Most places that
create these temporary wires just need them as a receptacle for connects
to plug stuff into, and the actual wire is not needed. These places now
create temporary `hw.wire`s that are guaranteed to be replaced with the
connected value at the end of the lowering.

The remaining two uses of temporary SV wires are inout ports of
instances for which we don't have a nice HW equivalent yet; and as an
lvalue for plusargs, which we can resolve later by making plusargs an
intrinsic at the HW level.

With this commit almost all `sv.wire`s are gone from LowerToHW output.